Rename library back to AetherX
