+++ AetherX/AESx64.py
@@ -1,5 +1,5 @@
-from str_arrows import string_with_arrows
-from AES_errors import AetherScriptError
+from functions import string_with_arrows
+from errors import AetherScriptError
 
 import string
 import os
